chore: v2_host test layout: unify on sibling tests/ directory [ORB-10387]
Agent-Model: claude-sonnet-5
