+++ crates/orbit-core/src/runtime/v2_host/tests/mod.rs
@@ -1,0 +1,9 @@
+mod backlog_exclusion;
+mod cli_executor;
+mod dispatch;
+mod learning_reminders;
+mod pipeline_actions;
+mod sandbox;
+mod task_context;
+mod triage;
+mod v2_host;
