+++ crates/orbit-core/src/runtime/v2_host/tests/triage.rs
@@ -12,7 +12,8 @@ use orbit_store::{JobRunStepParams, TaskCreateParams, TaskReservationReleaseReas
 use serde_json::{Value, json};
 use tempfile::tempdir;
 
-use super::*;
+use super::super::triage::*;
+use crate::OrbitRuntime;
 
 const PIPELINE_JOB: &str = "task_pr_pipeline";
 const ENVIRONMENTAL_STEP_MESSAGE: &str =
